Modified 1

+++ stats-test.cpp
@@ -5,7 +5,7 @@
 
 TEST(Statistics, ReportsAverageMinMax) {
     auto computedStats = Statistics::ComputeStatistics({1.5, 8.9, 3.2, 4.5});
-    float epsilon = 0.001;
+    double epsilon = 0.001;
     EXPECT_LT(std::abs(computedStats.average - 4.525), epsilon);
     EXPECT_LT(std::abs(computedStats.max - 8.9), epsilon);
     EXPECT_LT(std::abs(computedStats.min - 1.5), epsilon);
